wip: move project storage form to separate file
